requirements.txt
requests~=2.24.0
Pillow~=7.2.0